Save: 12/30/2023 6:17 PM

+++ Assets/#WindowDLLCode/2022_05_20_E_NativeWindowUtility/Runtime/Store/TDD/UserCoordinate/User32PercentPointRelative.cs
@@ -5,8 +5,8 @@ using UnityEngine;
 [System.Serializable]
 public class User32PercentPointRelative : IPercentPointGet
 {
-    public int m_xLeft2Right;
-    public int m_yTop2Down;
+    public float m_xLeft2Right;
+    public float m_yTop2Down;
 
     public void GetHorizontalDirection(out AxisDirectionHorizontal value)
     {
